# api/openapi.yaml
openapi: 3.0.3
info:
  title: ASM Platform API
  version: 0.8.0
  description: |
    Contract for the HTTP routes currently mounted by cmd/api. Discovery scope,
    task template, and task run management are implemented below service level
    but are not listed here until HTTP handlers are added.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /admin/users:
    get:
      tags: [platform-users]
      summary: List users in the current actor's tenant.
      description: Requires user:read. The tenant is always resolved from the authenticated actor.
      parameters:
        - $ref: "#/components/parameters/PageNumber"
        - $ref: "#/components/parameters/PageSize"
        - name: q
          in: query
          schema: { type: string, maxLength: 128 }
        - name: status
          in: query
          schema: { type: string, enum: [active, disabled] }
        - name: role
          in: query
          schema: { type: string, enum: [system_admin, security_admin, none] }
      responses:
        "200": { description: Tenant-scoped user page., content: { application/json: { schema: { $ref: "#/components/schemas/PlatformUserPageEnvelope" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [platform-users]
      summary: Create a user in the current actor's tenant and organization.
      description: Requires user:write. A non-null role additionally requires user:role:write. Usernames are normalized to lowercase ASCII and are immutable.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PlatformUserCreateRequest" }
      responses:
        "201": { description: Created user., content: { application/json: { schema: { $ref: "#/components/schemas/PlatformUserEnvelope" } } } }
        "400": { description: Malformed request., content: { application/json: { schema: { $ref: "#/components/schemas/ErrorEnvelope" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { description: Invalid profile, role, or temporary password., content: { application/json: { schema: { $ref: "#/components/schemas/ErrorEnvelope" } } } }
  /admin/users/{user_id}:
    get:
      tags: [platform-users]
      summary: Get a user in the current actor's tenant.
      parameters: [{ $ref: "#/components/parameters/UserID" }]
      responses:
        "200": { description: Platform user., content: { application/json: { schema: { $ref: "#/components/schemas/PlatformUserEnvelope" } } } }
        "400": { description: Invalid user id., content: { application/json: { schema: { $ref: "#/components/schemas/ErrorEnvelope" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { description: Missing or cross-tenant users both return not found., content: { application/json: { schema: { $ref: "#/components/schemas/ErrorEnvelope" } } } }
    patch:
      tags: [platform-users]
      summary: Update mutable platform-user profile fields.
      description: Username, status, password, and role are not mutable through this endpoint.
      parameters: [{ $ref: "#/components/parameters/UserID" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PlatformUserUpdateRequest" }
      responses:
        "200": { description: Updated user., content: { application/json: { schema: { $ref: "#/components/schemas/PlatformUserEnvelope" } } } }
        "400": { description: Invalid user id or malformed request., content: { application/json: { schema: { $ref: "#/components/schemas/ErrorEnvelope" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { description: Missing or cross-tenant user., content: { application/json: { schema: { $ref: "#/components/schemas/ErrorEnvelope" } } } }
        "422": { description: Invalid profile fields., content: { application/json: { schema: { $ref: "#/components/schemas/ErrorEnvelope" } } } }
  /admin/users/{user_id}/transitions:
    post:
      tags: [platform-users]
      summary: Enable or disable a platform user.
      description: Requires user:write. A reason is mandatory. Self-disable and removal of the last active system administrator are rejected. Successful transitions revoke existing sessions and are transactionally audited.
      parameters: [{ $ref: "#/components/parameters/UserID" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PlatformUserTransitionRequest" }
      responses:
        "200": { description: Updated user., content: { application/json: { schema: { $ref: "#/components/schemas/PlatformUserEnvelope" } } } }
        "400": { description: Invalid user id or malformed request., content: { application/json: { schema: { $ref: "#/components/schemas/ErrorEnvelope" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { description: Missing or cross-tenant user., content: { application/json: { schema: { $ref: "#/components/schemas/ErrorEnvelope" } } } }
        "422": { description: Invalid status or reason., content: { application/json: { schema: { $ref: "#/components/schemas/ErrorEnvelope" } } } }
        "409": { $ref: "#/components/responses/Conflict" }
  /admin/users/{user_id}/password-reset:
    post:
      tags: [platform-users]
      summary: Replace a user's password with an administrator-supplied temporary password.
      description: Requires user:credential:reset. The password is bcrypt-hashed, never returned or audited, and existing sessions are revoked.
      parameters: [{ $ref: "#/components/parameters/UserID" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PlatformUserPasswordResetRequest" }
      responses:
        "200": { description: Password reset acknowledgement., content: { application/json: { schema: { $ref: "#/components/schemas/GenericEnvelope" } } } }
        "400": { description: Invalid user id or malformed request., content: { application/json: { schema: { $ref: "#/components/schemas/ErrorEnvelope" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { description: Missing or cross-tenant user., content: { application/json: { schema: { $ref: "#/components/schemas/ErrorEnvelope" } } } }
        "422": { description: Temporary password violates policy., content: { application/json: { schema: { $ref: "#/components/schemas/ErrorEnvelope" } } } }
  /admin/users/{user_id}/tenant-role:
    put:
      tags: [platform-users]
      summary: Set or clear a user's tenant-level platform role.
      description: Requires user:role:write. A null role clears the assignment. Role changes revoke existing sessions and are transactionally audited.
      parameters: [{ $ref: "#/components/parameters/UserID" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TenantRoleUpdateRequest" }
      responses:
        "200": { description: Role assignment acknowledgement., content: { application/json: { schema: { $ref: "#/components/schemas/GenericEnvelope" } } } }
        "400": { description: Invalid user id or malformed request., content: { application/json: { schema: { $ref: "#/components/schemas/ErrorEnvelope" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { description: Missing or cross-tenant user., content: { application/json: { schema: { $ref: "#/components/schemas/ErrorEnvelope" } } } }
        "422": { description: Invalid tenant role., content: { application/json: { schema: { $ref: "#/components/schemas/ErrorEnvelope" } } } }
        "409": { $ref: "#/components/responses/Conflict" }
  /admin/users/{user_id}/projects:
    get:
      tags: [platform-users]
      summary: List a user's project memberships without modifying them.
      parameters: [{ $ref: "#/components/parameters/UserID" }]
      responses:
        "200": { description: Project membership list., content: { application/json: { schema: { $ref: "#/components/schemas/PlatformUserProjectsEnvelope" } } } }
        "400": { description: Invalid user id., content: { application/json: { schema: { $ref: "#/components/schemas/ErrorEnvelope" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { description: Missing or cross-tenant user., content: { application/json: { schema: { $ref: "#/components/schemas/ErrorEnvelope" } } } }
  /admin/roles:
    get:
      tags: [platform-users]
      summary: Return the fixed read-only role matrix.
      description: Requires user:read. Only system administrators receive user-management permissions.
      responses:
        "200": { description: Fixed role matrix., content: { application/json: { schema: { $ref: "#/components/schemas/AdminRolesEnvelope" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    UserID:
      name: user_id
      in: path
      required: true
      schema: { type: integer, format: uint64, minimum: 1 }
    PageNumber:
      name: page_number
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: page_size
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
  responses:
    Unauthorized:
      description: Authentication failed or missing.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Forbidden:
      description: The principal cannot access the requested tenant or project scope.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Conflict:
      description: State conflict.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
    GenericEnvelope:
      type: object
      required: [request_id, data]
      properties:
        request_id: { type: string }
        data: {}
    ErrorEnvelope:
      type: object
      required: [request_id, error]
      properties:
        request_id: { type: string }
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details: {}
...
    TenantRole:
      type: string
      nullable: true
      enum: [system_admin, security_admin]
    PlatformUser:
      type: object
      required: [id, username, name, email, phone, department, role, project_count, status, must_change_password, created_at, updated_at]
      properties:
        id: { type: integer, format: uint64 }
        username: { type: string, maxLength: 128, pattern: "^[a-z0-9][a-z0-9._-]{0,127}$" }
        name: { type: string, maxLength: 255 }
        email: { type: string, format: email, maxLength: 255 }
        phone: { type: string, maxLength: 32 }
        department: { type: string, maxLength: 128 }
        role: { $ref: "#/components/schemas/TenantRole" }
        project_count: { type: integer, format: int64, minimum: 0 }
        status: { type: string, enum: [active, disabled] }
        last_login_at: { type: string, format: date-time, nullable: true }
        must_change_password: { type: boolean }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    PlatformUserEnvelope:
      allOf:
        - $ref: "#/components/schemas/GenericEnvelope"
        - type: object
          properties:
            data: { $ref: "#/components/schemas/PlatformUser" }
    PlatformUserPageEnvelope:
      allOf:
        - $ref: "#/components/schemas/GenericEnvelope"
        - type: object
          properties:
            data:
              type: object
              required: [items, total, page_size, page_number]
              properties:
                items: { type: array, items: { $ref: "#/components/schemas/PlatformUser" } }
                total: { type: integer, format: int64, minimum: 0 }
                page_size: { type: integer, minimum: 1, maximum: 100 }
                page_number: { type: integer, minimum: 1 }
    PlatformUserCreateRequest:
      type: object
      required: [username, name, password]
      properties:
        username: { type: string, maxLength: 128, pattern: "^[A-Za-z0-9][A-Za-z0-9._-]{0,127}$" }
        name: { type: string, minLength: 1, maxLength: 255 }
        email: { type: string, format: email, maxLength: 255 }
        phone: { type: string, maxLength: 32 }
        department: { type: string, maxLength: 128 }
        role: { $ref: "#/components/schemas/TenantRole" }
        status: { type: string, enum: [active, disabled], default: active }
        password: { type: string, format: password, minLength: 12, maxLength: 72, writeOnly: true }
    PlatformUserUpdateRequest:
      type: object
      minProperties: 1
      additionalProperties: false
      properties:
        name: { type: string, minLength: 1, maxLength: 255 }
        email: { type: string, format: email, maxLength: 255 }
        phone: { type: string, maxLength: 32 }
        department: { type: string, maxLength: 128 }
    PlatformUserTransitionRequest:
      type: object
      required: [status, reason]
      additionalProperties: false
      properties:
        status: { type: string, enum: [active, disabled] }
        reason: { type: string, minLength: 1, maxLength: 512 }
    PlatformUserPasswordResetRequest:
      type: object
      required: [temporary_password]
      additionalProperties: false
      properties:
        temporary_password: { type: string, format: password, minLength: 12, maxLength: 72, writeOnly: true }
    TenantRoleUpdateRequest:
      type: object
      required: [role]
      additionalProperties: false
      properties:
        role: { $ref: "#/components/schemas/TenantRole" }
    PlatformUserProject:
      type: object
      required: [id, project_code, name, role, status, updated_at]
      properties:
        id: { type: integer, format: uint64 }
        project_code: { type: string }
        name: { type: string }
        role: { type: string, enum: [project_owner, security_ops, developer, viewer] }
        status: { type: string }
        updated_at: { type: string, format: date-time }
    PlatformUserProjectsEnvelope:
      allOf:
        - $ref: "#/components/schemas/GenericEnvelope"
        - type: object
          properties:
            data:
              type: object
              required: [items]
              properties:
                items: { type: array, items: { $ref: "#/components/schemas/PlatformUserProject" } }
    AdminRole:
      type: object
      required: [value, label, scope, permissions]
      properties:
        value: { type: string, enum: [system_admin, security_admin, project_owner, security_ops, developer, viewer] }
        label: { type: string }
        scope: { type: string, enum: [tenant, project] }
        permissions: { type: array, uniqueItems: true, items: { type: string } }
    AdminRolesEnvelope:
      allOf:
        - $ref: "#/components/schemas/GenericEnvelope"
        - type: object
          properties:
            data: { type: array, items: { $ref: "#/components/schemas/AdminRole" } }
...